refactor: address review feedback on bulk optimize filter
- Move filter docblock next to the apply_filters call
- Use wpm_apply_filters_typed() instead of discouraged apply_filters()
- Fix PHPStan withArgs() error via with() + Mockery::on()
- Extract BulkWithThreeOptimizeIdsStub to its own file
