+++ Tests/Unit/classes/Bulk/RunOptimizeTest.php
@@ -6,8 +6,8 @@ namespace Imagify\Tests\Unit\classes\Bulk;
 use Brain\Monkey\Filters;
 use Brain\Monkey\Functions;
 use Imagify\Bulk\Bulk;
-use Imagify\Bulk\BulkInterface;
 use Imagify\Tests\Unit\TestCase;
+use Imagify\Tests\Unit\classes\Bulk\Stubs\BulkWithThreeOptimizeIdsStub;
 use Mockery;
 
 /**
@@ -62,14 +62,16 @@ class RunOptimizeTest extends TestCase {
 
 		Functions\expect( 'as_enqueue_async_action' )
 			->twice()
-			->withArgs(
-				function ( $hook, $args, $group ) {
-					return 'imagify_optimize_media' === $hook
-						&& in_array( $args['id'], [ 1, 3 ], true )
-						&& 'wp' === $args['context']
-						&& 1 === $args['level']
-						&& 'imagify-wp-optimize-media' === $group;
-				}
+			->with(
+				'imagify_optimize_media',
+				Mockery::on(
+					function ( $args ) {
+						return in_array( $args['id'], [ 1, 3 ], true )
+							&& 'wp' === $args['context']
+							&& 1 === $args['level'];
+					}
+				),
+				'imagify-wp-optimize-media'
 			);
 
 		Functions\expect( 'set_transient' )
@@ -122,52 +124,5 @@ class RunOptimizeTest extends TestCase {
 			],
 			$result
 		);
-	}
-}
-
-/**
- * Stub bulk: three eligible media IDs for optimization.
- */
-class BulkWithThreeOptimizeIdsStub implements BulkInterface {
-
-	/**
-	 * {@inheritdoc}
-	 */
-	public function get_unoptimized_media_ids( $optimization_level ) {
-		return [ 1, 2, 3 ];
-	}
-
-	/**
-	 * {@inheritdoc}
-	 */
-	public function get_optimized_media_ids(): array {
-		return [];
-	}
-
-	/**
-	 * {@inheritdoc}
-	 */
-	public function get_optimized_media_ids_without_format( $format ) {
-		return [
-			'ids'    => [],
-			'errors' => [
-				'no_file_path' => [],
-				'no_backup'    => [],
-			],
-		];
-	}
-
-	/**
-	 * {@inheritdoc}
-	 */
-	public function has_optimized_media_without_nextgen() {
-		return 0;
-	}
-
-	/**
-	 * {@inheritdoc}
-	 */
-	public function get_context_data() {
-		return [];
 	}
 }
